third_party/blink/web_tests/wpt_internal/attribution-reporting/navigation-source-trigger-data-cardinality-test.sub.https.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
const navigationSourceTriggerDataCardinality = 8;

const trigger_data_cardinality_promise_test = (value) =>
  attribution_reporting_promise_test(async t => {
    await registerAttributionSrcWithOpen('Attribution-Reporting-Register-Source', {
      source_event_id: '999',
      destination: `https://{{host}}`,
    });

    registerAttributionSrc('Attribution-Reporting-Register-Trigger', {
      event_trigger_data: [{trigger_data: `${value}`}],
    });

    const payload = await pollEventLevelReports();
    assert_equals(payload.reports.length, 1);
    const report = JSON.parse(payload.reports[0].body);

    const expected = value % navigationSourceTriggerDataCardinality;
    assert_equals(report.trigger_data, `${expected}`);
  }, `Trigger data ${value} is sanitized according to source type.`);

for (let i = 0; i < navigationSourceTriggerDataCardinality * 2; i++) {
  trigger_data_cardinality_promise_test(i);
}
</script>
